+++ src/Methods/FasterGSFusedRapid/FasterGSFusedRapidCudaBackend/FasterGSFusedRapidCudaBackend/rasterization/include/forward.h
@@ -22,7 +22,6 @@ namespace faster_gs::rasterization {
         const float3* cam_position,
         const float3* bg_color,
         float* image,
-        float* inv_depth,
         float* metric_counts,
         const int n_primitives,
         const int active_sh_bases,